Move MapEntryAssertionSamples to deprecated package
