+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/samples/deprecated/MapEntryAssertionSamples.kt
@@ -1,6 +1,7 @@
-package ch.tutteli.atrium.api.fluent.en_GB.samples
+package ch.tutteli.atrium.api.fluent.en_GB.samples.deprecated
 
 import ch.tutteli.atrium.api.fluent.en_GB.*
+import ch.tutteli.atrium.api.fluent.en_GB.samples.fails
 import ch.tutteli.atrium.api.verbs.internal.expect
 import kotlin.test.Test
